Move code to it's core functionalities so the main app is more clean
