+++ indexer/index-manager/src/config_helper.rs
@@ -10,6 +10,20 @@ lazy_static! {
         env::var("IPFS_ADDRESS").unwrap_or(String::from("0.0.0.0:5001"));
 }
 
+// pub async fn get_index_config(ipfs_config_hash: &String) -> serde_yaml::Mapping {
+//     let ipfs_addresses = vec![IPFS_ADDRESS.to_string()];
+//     let ipfs_clients = create_ipfs_clients(&ipfs_addresses).await; // Refactor to use lazy load
+//
+//     let file_bytes = ipfs_clients[0]
+//         .cat_all(ipfs_config_hash.to_string())
+//         .compat()
+//         .await
+//         .unwrap()
+//         .to_vec();
+//
+//     serde_yaml::from_slice(&file_bytes).unwrap()
+// }
+
 pub async fn get_raw_query_from_ipfs(ipfs_model_hash: &String) -> String {
     log::info!("[Index Manager Helper] Downloading Raw Query from IPFS");
     let ipfs_addresses = vec![IPFS_ADDRESS.to_string()];
@@ -103,4 +117,14 @@ pub fn get_config_file_from_local(config_path: &String) -> String {
 pub fn get_mapping_file_from_local(mapping_path: &String) -> PathBuf {
     let so_file_path = PathBuf::from(mapping_path.to_string());
     so_file_path
+}
+
+
+pub fn read_config_file(config_file_path: &String) -> serde_yaml::Value {
+    let mut project_config_string = String::new();
+    let mut f = File::open(config_file_path).expect("Unable to open file"); // Refactor: Config to download config file from IPFS instead of just reading from local
+    f.read_to_string(&mut project_config_string)
+        .expect("Unable to read string"); // Get raw query
+    let project_config: serde_yaml::Value = serde_yaml::from_str(&project_config_string).unwrap();
+    project_config
 }
