+++ indexer/index-manager/src/lib.rs
@@ -3,3 +3,5 @@ pub mod types;
 pub mod helper;
 pub mod builder;
 pub mod config_helper;
+pub mod hasura;
+pub mod store;
